backend/requirements.txt
langgraph>=1.0.0
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-core>=1.0.0
langchain-mcp-adapters>=0.1.0
mcp[cli]>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pyyaml>=6.0
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
